+++ CHANGELOG.md
@@ -8,8 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Added `image.created` as a valid hub notification
-- 'manual' as a valid container deployment type
+- `manual` and `ephemeral` are now valid container deployment type
 - `Containers.Servers.usable()` call to get servers that this container COULD be deployed to
+- `Containers.Instances.create()` and `Containers.Instances.remove()` for containers with deployment strategy of `manual`
 
 ### Fixed